chore: move api error from client
