+++ apps/frontend/src/shared/api/client.ts
@@ -1,33 +1,6 @@
+import { ApiError } from "./errors";
+
 const TOKEN_KEY = "auth_token";
-
-export class ApiError extends Error {
-  code: string;
-  status: number;
-  statusText: string;
-  body: {
-    error: string;
-    code: string;
-    status: number;
-    details?: unknown;
-  };
-
-  constructor(
-    status: number,
-    statusText: string,
-    body: {
-      error: string;
-      code: string;
-      status: number;
-      details?: unknown;
-    },
-  ) {
-    super(body.error || `API Error ${status}: ${statusText}`);
-    this.code = body.code || "UNKNOWN_ERROR";
-    this.status = status;
-    this.statusText = statusText;
-    this.body = body;
-  }
-}
 
 export function getToken(): string | null {
   return localStorage.getItem(TOKEN_KEY);
